requirements.txt
clickhouse-connect>=0.7.19
python-dotenv>=1.0.1
websockets>=12.0
certifi>=2024.2.2
sqlparse>=0.5.0
fastapi>=0.111
uvicorn[standard]>=0.30